refactor(linear,github): split IssueSessionRepo per provider — strict isolation
Until now `linear_issue_sessions` was the only per-issue session table; both
LinearProvider and GitHubProvider read/wrote it via a shared
`D1IssueSessionRepo` instance + shared `IssueSessionRepo` interface. The
table name lied (rows of `<repo>#<number>` GitHub keys lived alongside
Linear UUIDs), schema changes for one provider tied to the other, and the
shared interface forced the two providers to pretend they had identical
needs (they don't — Linear has PAT-mode `claim`, GitHub doesn't).

This PR splits storage, classes, and interfaces. No backwards compat.

NEW STORAGE
- 0005_github_issue_sessions.sql migration creates `github_issue_sessions`
  (mirror DDL of `linear_issue_sessions`)
- packages/schema bootstrap DDL adds the same table for fresh deploys

NEW INTERFACES (each provider owns its own)
- LinearIssueSessionRepo — packages/linear/src/ports.ts. Full method set
  including `claim` (PAT-mode autopilot sweep) + webhook claim methods.
- GitHubIssueSessionRepo — packages/github/src/ports.ts. Webhook claim
  methods only (no `claim` — GitHub Apps are the only install path).
- IssueSessionRepo / IssueSession / IssueSessionStatus removed from
  integrations-core. `Container.issueSessions` slot deleted.

NEW CLASSES
- D1LinearIssueSessionRepo — d1/linear/issue-session-repo.ts (moved from
  d1/issue-session-repo.ts, hardcodes `linear_issue_sessions`)
- D1GitHubIssueSessionRepo — d1/github/issue-session-repo.ts (new,
  hardcodes `github_issue_sessions`)
- Same Linear/GitHub split for the node SqlClient adapter
- No shared parent or template — full isolation

PROVIDER WIRING
- LinearContainer extends Container with `linearIssueSessions` slot
- GitHubContainer extends Container with `githubIssueSessions` slot
- LinearProvider: this.container.issueSessions → linearIssueSessions
- GitHubProvider: this.container.issueSessions → githubIssueSessions
- cf-container / node-container construct two repo instances

TEST FAKES
- packages/linear/src/test-fakes.ts — InMemoryLinearIssueSessionRepo +
  buildFakeLinearContainer
- packages/github/src/test-fakes.ts — adds InMemoryGitHubIssueSessionRepo
  to existing fakes; buildFakeGitHubContainer wires it
- InMemoryIssueSessionRepo dropped from integrations-core test-fakes

Existing GitHub data on staging in `linear_issue_sessions` becomes orphan;
acceptable per "no backwards compat" directive (staging-only test data).
