+++ packages/integrations-adapters-cf/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@open-managed-agents/github": "workspace:*",
     "@open-managed-agents/integrations-core": "workspace:*",
+    "@open-managed-agents/linear": "workspace:*",
     "@open-managed-agents/shared": "workspace:*",
     "@open-managed-agents/slack": "workspace:*"
   }
