+++ packages/integrations-adapters-cf/src/index.ts
@@ -22,7 +22,8 @@ export { D1GitHubInstallationRepo } from "./d1/github/installation-repo";
 export { D1GitHubPublicationRepo } from "./d1/github/publication-repo";
 export { D1GitHubWebhookEventStore } from "./d1/github/webhook-event-store";
 export { D1LinearEventStore } from "./d1/linear-event-store";
-export { D1IssueSessionRepo } from "./d1/issue-session-repo";
+export { D1LinearIssueSessionRepo } from "./d1/linear/issue-session-repo";
+export { D1GitHubIssueSessionRepo } from "./d1/github/issue-session-repo";
 export { D1SetupLinkRepo } from "./d1/setup-link-repo";
 export { D1DispatchRuleRepo } from "./d1/dispatch-rule-repo";
 export { D1TenantResolver } from "./d1/tenant-resolver";
