+++ packages/integrations-adapters-node/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@open-managed-agents/github": "workspace:*",
     "@open-managed-agents/integrations-core": "workspace:*",
+    "@open-managed-agents/linear": "workspace:*",
     "@open-managed-agents/sql-client": "workspace:*",
     "@open-managed-agents/shared": "workspace:*",
     "@open-managed-agents/slack": "workspace:*"
