+++ packages/integrations-adapters-node/src/index.ts
@@ -25,7 +25,8 @@ export { SqlGitHubInstallationRepo } from "./sql/github/installation-repo";
 export { SqlGitHubPublicationRepo } from "./sql/github/publication-repo";
 export { SqlGitHubWebhookEventStore } from "./sql/github/webhook-event-store";
 export { SqlLinearEventStore } from "./sql/linear-event-store";
-export { SqlIssueSessionRepo } from "./sql/issue-session-repo";
+export { SqlLinearIssueSessionRepo } from "./sql/linear/issue-session-repo";
+export { SqlGitHubIssueSessionRepo } from "./sql/github/issue-session-repo";
 export { SqlSetupLinkRepo } from "./sql/setup-link-repo";
 export { SqlDispatchRuleRepo } from "./sql/dispatch-rule-repo";
 export { SqlMembershipTenantResolver } from "./sql/membership-tenant-resolver";
