+++ packages/integrations-adapters-node/src/node-container.ts
@@ -37,7 +37,8 @@ import { SqlGitHubInstallationRepo } from "./sql/github/installation-repo";
 import { SqlGitHubPublicationRepo } from "./sql/github/publication-repo";
 import { SqlGitHubWebhookEventStore } from "./sql/github/webhook-event-store";
 import { SqlInstallationRepo } from "./sql/installation-repo";
-import { SqlIssueSessionRepo } from "./sql/issue-session-repo";
+import { SqlLinearIssueSessionRepo } from "./sql/linear/issue-session-repo";
+import { SqlGitHubIssueSessionRepo } from "./sql/github/issue-session-repo";
 import { SqlLinearEventStore } from "./sql/linear-event-store";
 import { SqlPublicationRepo } from "./sql/publication-repo";
 import { SqlSetupLinkRepo } from "./sql/setup-link-repo";
@@ -77,7 +78,8 @@ export function buildNodeRepos(env: NodeReposEnv) {
   const githubApps = new SqlGitHubAppRepo(sql, cryptoImpl, ids);
   const linearEvents = new SqlLinearEventStore(sql);
   const githubWebhookEvents = new SqlGitHubWebhookEventStore(sql);
-  const issueSessions = new SqlIssueSessionRepo(sql);
+  const linearIssueSessions = new SqlLinearIssueSessionRepo(sql);
+  const githubIssueSessions = new SqlGitHubIssueSessionRepo(sql);
   const setupLinks = new SqlSetupLinkRepo(sql, ids);
   const dispatchRules = new SqlDispatchRuleRepo(sql, ids);
   const sessionScopes = new SqlSlackSessionScopeRepo(sql);
@@ -98,7 +100,8 @@ export function buildNodeRepos(env: NodeReposEnv) {
     githubApps,
     linearEvents,
     githubWebhookEvents,
-    issueSessions,
+    linearIssueSessions,
+    githubIssueSessions,
     sessionScopes,
     setupLinks,
     dispatchRules,
